Moved Dashboard client-side components out
